Implement List<EntityType> fetchProductsFromIds(List<EntityType> emptyProducts) for ProductRepository, CustomerRepository, SalesInvoiceRepository, ServiceRepository, StockTransactionRepository. These function accepts a List of entities with all null fields but Id. Finds existing entity with this Id and creates a List of existing entities.
This allow client to send only id field of related entity when building relationship between two entities.

Example:

POST /customer/
{ "name": "someCustomer"}
which returns a Customer with "id": 0

so when creating a SalesInvoice we can
{
     "customer": {"id":0},
     "documentNumber": "AAA1212CCC",
}
which automatically puts the Customer with "name": "someCustomer" to customer field of created SalesInvoice entity. The benefit is we didn't need to provide all details of Customer entity.

+++ src/main/java/com/logo/model/ProductOrService.java
@@ -8,7 +8,7 @@ import java.util.Currency;
 
 @Data
 abstract class ProductOrService {
-    protected long id;
+    protected int id;
     protected String name;
      UnitType unitType;
     //KDV(Katma Değer Vergisi)/VAT(Value Added Tax)
